Move VLQ encoding into a separate module
